More 'just for fun': Plot of a more complete set of counties with unambiguous SLCSP
In the previous versino of the plot, we processed each zip code in zips.csv, computed
the results, and then plotted the result for the county corresponding to that zip code.
This is more work than necessary (potentially many zips in one county) and also produces
less than the complete set of possible results (when there's more than one rate area in
the zip code, it gets marked as ambiguous).

This version simply computes the answer for each listed FIPS county code.
It still leaves some counties without results for a few valid reasons (e.g. if they
don't have more than one silver plan), but the set is more complete and runs faster.